services/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.9.1
pydantic-settings==2.4.0
pandas==2.2.3
numpy==2.1.2
python-slugify==8.0.4
scikit-learn==1.5.2
httpx==0.27.2
torch==2.5.1+cpu
requests>=2.31.0
beautifulsoup4>=4.12.0